+++ lib/karafka/web/ui/pro/controllers/health.rb
@@ -21,7 +21,12 @@ module Karafka
             # Displays the current system state
             def overview
               current_state = Models::ConsumersState.current!
-              @stats = refine(Models::Health.current(current_state))
+              @stats = Models::Health.current(current_state)
+
+              # Refine only on a per topic basis not to resort higher levels
+              @stats.each_value do |cg_details|
+                cg_details.each_value { |topic_details| refine(topic_details) }
+              end
 
               render
             end